Feature/refatoracao fluxo

+++ src/main/java/br/com/nlw/events/application/usecases/subscription/gateway/CreateSubscriptionUseCase.java
@@ -1,10 +1,8 @@
 package br.com.nlw.events.application.usecases.subscription.gateway;
 
-import br.com.nlw.events.domain.model.Subscription;
-import br.com.nlw.events.domain.model.User;
-
-import java.util.UUID;
+import br.com.nlw.events.domain.models.Subscription;
+import br.com.nlw.events.domain.models.User;
 
 public interface CreateSubscriptionUseCase {
-    Subscription execute(final String prettyName, final Integer userIndicatorId, final User userRequest);
+    Subscription execute(final String prettyName, final Long userIndicatorId, final User userRequest);
 }
